requirements.txt
google-generativeai
python-dotenv
fastapi
uvicorn
python-multipart
python-docx
markdown
fpdf2
pandas
openpyxl
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose
bcrypt==4.0.1
google-auth
google-api-python-client
